requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
anthropic>=0.40.0
python-dotenv==1.0.0
requests
flask==3.0.0
flask-session==0.5.0
